requirements.txt
amqp==1.4.9
redis==2.10.5
anyjson==0.3.3
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
billiard==3.3.0.23
bitarray==0.8.1
celery==3.1.24
click==6.6
decorator==4.0.10
enum34==1.1.6
Flask==0.11.1
Flask-Migrate==2.0.2
thrift_sasl==0.2.1
Flask-Cors==3.0.2
impyla==0.13.8
inflection==0.3.1
ipython==5.1.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.37
MarkupSafe==0.23
marshmallow==2.10.3
nose==1.3.7
numpy
pandas
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.7
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.1.3
python-dateutil==2.5.3
pytz==2016.7
PyYAML==3.12
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.3
thrift==0.9.3
traitlets==4.3.1
vine==1.1.3
wcwidth==0.1.7
Werkzeug==0.11.11
boto3==1.4.2
requests==2.12.4
ddt==1.1.1